requirements.txt
flake8
docutils
google-cloud-vision
